Automatic Semicolon Insertion
If we forget to insert semicolon the javascript automatically inserts for us. But this will be a problem as illustrated in the code with the return statement where it inserts semicolon and leads to undefined. which is why we have to place the curly braces of the function in the same line, which prevents this type of problems