Closures Eg-2
In the enclosed code first in starter it outputs 3 for all as I will be referenced in the outer environment. This can be overcome as illustrated in finished file by using a new variable j and initialising it to i which creates a new execution environment and then immediately invoking the function.